Create sticks
Problem C: Square
Given a set of sticks of various lengths, is it possible to join them end-to-end to form a square?

The first line of input contains N, the number of test cases. Each test case begins with an integer 4 <= M <= 20, the number of sticks. M integers follow; each gives the length of a stick - an integer between 1 and 10,000.

For each case, output a line containing "yes" if is is possible to form a square; otherwise output "no".

Samples
input 
3
4 1 1 1 1
5 10 20 30 40 50
8 1 7 2 6 4 4 3 5
output 
yes
no
yes